% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Dummy function}
\usage{
dummy(x)
}
\arguments{
\item{x}{Text to be displayed}

\item{return}{The function has no return value}
}
\description{
This is a dummy function to be used as a template - with imports
}
\examples{
dummy("I am a dummy function call!")
}
