% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Methods}
\alias{Methods}
\alias{Math.ref}
\alias{Ops.ref}
\alias{Complex.ref}
\alias{Summary.ref}
\alias{Math.rfexpr}
\alias{Ops.rfexpr}
\alias{Complex.rfexpr}
\alias{Summary.rfexpr}
\alias{all.equal.ref}
\alias{anyDuplicated.ref}
\alias{as.character.ref}
\alias{as.data.frame.ref}
\alias{as.Date.ref}
\alias{as.double.ref}
\alias{as.function.ref}
\alias{as.list.ref}
\alias{as.matrix.ref}
\alias{as.POSIXct.ref}
\alias{as.POSIXlt.ref}
\alias{as.single.ref}
\alias{as.table.ref}
\alias{c.ref}
\alias{cut.ref}
\alias{diff.ref}
\alias{dim.ref}
\alias{droplevels.ref}
\alias{duplicated.ref}
\alias{format.ref}
\alias{isSymmetric.ref}
\alias{kappa.ref}
\alias{labels.ref}
\alias{length.ref}
\alias{levels.ref}
\alias{mean.ref}
\alias{merge.ref}
\alias{qr.ref}
\alias{rep.ref}
\alias{rev.ref}
\alias{round.ref}
\alias{row.names.ref}
\alias{solve.ref}
\alias{sort.ref}
\alias{aggregate.ref}
\alias{coef.ref}
\alias{confint.ref}
\alias{fitted.ref}
\alias{median.ref}
\alias{model.frame.ref}
\alias{model.matrix.ref}
\alias{na.omit.ref}
\alias{plot.ref}
\alias{predict.ref}
\alias{residuals.ref}
\alias{summary.ref}
\alias{terms.ref}
\alias{vcov.ref}
\alias{window.ref}
\alias{all.equal.rfexpr}
\alias{anyDuplicated.rfexpr}
\alias{as.character.rfexpr}
\alias{as.data.frame.rfexpr}
\alias{as.Date.rfexpr}
\alias{as.double.rfexpr}
\alias{as.function.rfexpr}
\alias{as.list.rfexpr}
\alias{as.matrix.rfexpr}
\alias{as.POSIXct.rfexpr}
\alias{as.POSIXlt.rfexpr}
\alias{as.single.rfexpr}
\alias{as.table.rfexpr}
\alias{c.rfexpr}
\alias{cut.rfexpr}
\alias{diff.rfexpr}
\alias{dim.rfexpr}
\alias{droplevels.rfexpr}
\alias{duplicated.rfexpr}
\alias{format.rfexpr}
\alias{isSymmetric.rfexpr}
\alias{kappa.rfexpr}
\alias{labels.rfexpr}
\alias{length.rfexpr}
\alias{levels.rfexpr}
\alias{mean.rfexpr}
\alias{merge.rfexpr}
\alias{qr.rfexpr}
\alias{rep.rfexpr}
\alias{rev.rfexpr}
\alias{round.rfexpr}
\alias{row.names.rfexpr}
\alias{solve.rfexpr}
\alias{sort.rfexpr}
\alias{aggregate.rfexpr}
\alias{coef.rfexpr}
\alias{confint.rfexpr}
\alias{fitted.rfexpr}
\alias{median.rfexpr}
\alias{model.frame.rfexpr}
\alias{model.matrix.rfexpr}
\alias{na.omit.rfexpr}
\alias{plot.rfexpr}
\alias{predict.rfexpr}
\alias{residuals.rfexpr}
\alias{summary.rfexpr}
\alias{terms.rfexpr}
\alias{vcov.rfexpr}
\alias{window.rfexpr}
\title{S3 Methods for Automatic Dereferencing}
\usage{
\method{Math}{ref}(x, ...)

\method{Ops}{ref}(e1, e2)

\method{Complex}{ref}(z)

\method{Summary}{ref}(..., na.rm = FALSE)

\method{Math}{rfexpr}(x, ...)

\method{Ops}{rfexpr}(e1, e2)

\method{Complex}{rfexpr}(z)

\method{Summary}{rfexpr}(..., na.rm = FALSE)

\method{all.equal}{ref}(target, current, ...)

\method{anyDuplicated}{ref}(x, incomparables = FALSE, ...)

\method{as.character}{ref}(x, ...)

\method{as.data.frame}{ref}(x, row.names = NULL, optional = FALSE, ...)

\method{as.Date}{ref}(x, ...)

\method{as.double}{ref}(x, ...)

\method{as.function}{ref}(x, ...)

\method{as.list}{ref}(x, ...)

\method{as.matrix}{ref}(x, ...)

\method{as.POSIXct}{ref}(x, tz = "", ...)

\method{as.POSIXlt}{ref}(x, tz = "", ...)

\method{as.single}{ref}(x, ...)

\method{as.table}{ref}(x, ...)

\method{c}{ref}(...)

\method{cut}{ref}(x, ...)

\method{diff}{ref}(x, ...)

\method{dim}{ref}(x)

\method{droplevels}{ref}(x, ...)

\method{duplicated}{ref}(x, incomparables = FALSE, ...)

\method{format}{ref}(x, ...)

\method{isSymmetric}{ref}(object, ...)

\method{kappa}{ref}(z, ...)

\method{labels}{ref}(object, ...)

\method{length}{ref}(x)

\method{levels}{ref}(x)

\method{mean}{ref}(x, ...)

\method{merge}{ref}(x, y, ...)

\method{qr}{ref}(x, ...)

\method{rep}{ref}(x, ...)

\method{rev}{ref}(x)

\method{round}{ref}(x, digits = 0)

\method{row.names}{ref}(x)

\method{solve}{ref}(a, b, ...)

\method{sort}{ref}(x, decreasing = FALSE, ...)

\method{aggregate}{ref}(x, ...)

\method{coef}{ref}(object, ...)

\method{confint}{ref}(object, parm, level = 0.95, ...)

\method{fitted}{ref}(object, ...)

\method{median}{ref}(x, na.rm = FALSE, ...)

\method{model.frame}{ref}(formula, ...)

\method{model.matrix}{ref}(object, ...)

\method{na.omit}{ref}(object, ...)

\method{plot}{ref}(x, y, ...)

\method{predict}{ref}(object, ...)

\method{residuals}{ref}(object, ...)

\method{summary}{ref}(object, ...)

\method{terms}{ref}(x, ...)

\method{vcov}{ref}(object, ...)

\method{window}{ref}(x, ...)

\method{all.equal}{rfexpr}(target, current, ...)

\method{anyDuplicated}{rfexpr}(x, incomparables = FALSE, ...)

\method{as.character}{rfexpr}(x, ...)

\method{as.data.frame}{rfexpr}(x, row.names = NULL, optional = FALSE, ...)

\method{as.Date}{rfexpr}(x, ...)

\method{as.double}{rfexpr}(x, ...)

\method{as.function}{rfexpr}(x, ...)

\method{as.list}{rfexpr}(x, ...)

\method{as.matrix}{rfexpr}(x, ...)

\method{as.POSIXct}{rfexpr}(x, tz = "", ...)

\method{as.POSIXlt}{rfexpr}(x, tz = "", ...)

\method{as.single}{rfexpr}(x, ...)

\method{as.table}{rfexpr}(x, ...)

\method{c}{rfexpr}(...)

\method{cut}{rfexpr}(x, ...)

\method{diff}{rfexpr}(x, ...)

\method{dim}{rfexpr}(x)

\method{droplevels}{rfexpr}(x, ...)

\method{duplicated}{rfexpr}(x, incomparables = FALSE, ...)

\method{format}{rfexpr}(x, ...)

\method{isSymmetric}{rfexpr}(object, ...)

\method{kappa}{rfexpr}(z, ...)

\method{labels}{rfexpr}(object, ...)

\method{length}{rfexpr}(x)

\method{levels}{rfexpr}(x)

\method{mean}{rfexpr}(x, ...)

\method{merge}{rfexpr}(x, y, ...)

\method{qr}{rfexpr}(x, ...)

\method{rep}{rfexpr}(x, ...)

\method{rev}{rfexpr}(x)

\method{round}{rfexpr}(x, digits = 0)

\method{row.names}{rfexpr}(x)

\method{solve}{rfexpr}(a, b, ...)

\method{sort}{rfexpr}(x, decreasing = FALSE, ...)

\method{aggregate}{rfexpr}(x, ...)

\method{coef}{rfexpr}(object, ...)

\method{confint}{rfexpr}(object, parm, level = 0.95, ...)

\method{fitted}{rfexpr}(object, ...)

\method{median}{rfexpr}(x, na.rm = FALSE, ...)

\method{model.frame}{rfexpr}(formula, ...)

\method{model.matrix}{rfexpr}(object, ...)

\method{na.omit}{rfexpr}(object, ...)

\method{plot}{rfexpr}(x, y, ...)

\method{predict}{rfexpr}(object, ...)

\method{residuals}{rfexpr}(object, ...)

\method{summary}{rfexpr}(object, ...)

\method{terms}{rfexpr}(x, ...)

\method{vcov}{rfexpr}(object, ...)

\method{window}{rfexpr}(x, ...)
}
\arguments{
\item{x, y, e1, e2, z, target, current, object, a, b, formula}{objects of class \code{"ref"}}

\item{...}{other objects passed to the function}

\item{incomparables, digits, tz, row.names, optional, decreasing, na.rm, parm, level}{function specific arguments. See the relevant
functions for more details}
}
\value{
An R object depending on the function.
}
\description{
These functions automatically call \code{\link{deref}} when applied to a \code{\link{ref}} or \code{"rfexpr"}
object. Therefore, there is no need to explicitly call \code{deref}. \code{\link{sref}} objects will need
to be explicitly dereferenced before applying these functions. All functions are from \code{base} R.
}
