% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.R
\name{ref_list}
\alias{ref_list}
\title{Create A List of References}
\usage{
ref_list(...)
}
\arguments{
\item{...}{objects to be referenced, possibly named.}
}
\value{
a list containing object references
}
\description{
Create a list of references or referenced expressions. See \code{\link{ref}} for more details.
}
\examples{
x <- 1
y <- "hello"
z <- list(a = 1, b = 2, c = 3)

new_list <- ref_list(x, second = y, z)

!new_list[[1]]
(!new_list$second) == y  # TRUE

y <- 18
(!new_list$second) == 18 # TRUE

}
