% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{refer}
\alias{refer}
\title{Reference list from data objects}
\description{
The purpose of this extension is to facilitate the extraction of reference information from data objects that are based on published literature. Functions are implemented to prepare a list of source literature that comprises the data objects. This R extension was supported by the FAU GeoZentrum Nordbayern and is developed under the umbrella of the DFG (Deutsche Forschungsgemeinschaft) Research Unit TERSANE2 (For 2332, TEmperature Related Stressors in ANcient Extinctions).
}
\details{
This is still the Beta version. As is R, this is free software and comes with ABSOLUTELY NO WARRANTY. Nevertheless, notes about found bugs and suggestions are more than welcome.
}
\author{
Nussaibah B. Raja (nussaibahraja@gmail.com) Adam T. Kocsis
}
