% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbib.R
\name{subset_bib}
\alias{subset_bib}
\title{Subset references according to an index}
\usage{
subset_bib(fname, index, output)
}
\arguments{
\item{fname}{The name of the file containing the references. Should be a bib file.}

\item{index}{Vector of indices to be subset}

\item{output}{the name of the output file}
}
\value{

}
\description{
Subset references according to an index
}
\examples{
\dontrun{
index <- c(61274, 30858, 65677)
subset_bib("pbdb_references.bib", index, "references.bib")
}
}
