% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{wrap_url}
\alias{wrap_url}
\title{Breaks a url over multiple lines so it fits on an A4 page}
\usage{
wrap_url(x, width = 50, by = " ")
}
\arguments{
\item{x}{a vector containing the urls}

\item{width}{the maximum number of characters in one line}

\item{by}{by which character to break the lines}
}
\description{
Breaks a url over multiple lines so it fits on an A4 page
}
